move some function to lib
