+++ lib/index.js
@@ -1,18 +1,6 @@
-'use strict';
-
-const koa = require('koa');
-const koaBody = require('koa-body');
-const GithubWebhookHandler = require('koa-github-webhook-handler');
-const github = require('./lib/github');
+const github = require('./github');
 const bumpVersion = require('semver-increment');
 const promisify = require('es6-promisify');
-
-const app = koa();
-
-const githubWebhookHandler = new GithubWebhookHandler({
-  path: '/webhook',
-  secret: process.env.SEANBOT_WEBHOOK_SECRET,
-});
 
 const getPullRequests = promisify(github.pullRequests.get.bind(github.pullRequests));
 const getContent = promisify(github.repos.getContent.bind(github.repos));
@@ -98,17 +86,6 @@ function issueCommentHandler(event) {
   .then((res) => console.log(res))
   .catch((err) => console.error(err.message && err.stack));
 }
-
-githubWebhookHandler.on('issue_comment', issueCommentHandler);
-
-app.use(koaBody({
-  formidable: {
-    uploadDir: __dirname,
-  },
-}));
-
-app.use(githubWebhookHandler.middleware());
-app.listen((process.env.PORT || 4444));
 
 exports.getMasksFromCommentBody = getMasksFromCommentBody;
 exports.getMsgFormPrUrl = getMsgFormPrUrl;
