+++ package.json
@@ -19,7 +19,7 @@
     "semver-increment": "^0.1.0"
   },
   "devDependencies": {
-    "tap": "^5.7.1"
+    "tap": "^5.7.1",
     "babel-eslint": "^6.0.3",
     "eslint": "^2.8.0",
     "eslint-config-weflex": "^1.0.3",
